sculptor/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sculptor"

# This is going to be the true version.
version = "0.5.0"

readme = "README.md"
dependencies = [
    # Please add dev dependencies to dependency-groups.dev, not here.
    # Please keep these sorted alphabetically.
    "alembic>=1.16.1",
    "anthropic>=0.38.0",
    "beautifulsoup4>=4.12.2",
    "coolname>=2.2.0",
    "dockerfile-parse>=2.0.1",
    "email-validator",
    "fastapi",
    "filelock>=3.8.0",
    "humanfriendly>=10.0",
    "json5>=0.9.0",
    "loguru",
    "modal>=1.0.3",
    "psutil>=5.9.0",
    "psycopg[binary]",
    "pydantic-settings",
    "pyjwt[crypto]",
    "requests>=2.28.0",
    "sentry-sdk",
    "splinter>=0.19.0",
    "sqlalchemy",
    "tomlkit>=0.13.0",
    "typeid-python",
    "typer",
    "uvicorn>=0.34.3",
    "watchdog>=6.0.0",
    "websockets>=15.0.1",

    # There's a gap and then last of all, because of the non-standard way we import it, we have
    "imbue-core",

    # Once again, sort your import dependency!
]
requires-python = ">=3.11"

[tool.uv.sources]
imbue-core = { workspace = true }

[tool.setuptools.packages.find]
exclude = [
    "keys",

    # Exclude any testing packages from being included in our code.
    "tests",
    "tests.*",
    "*.tests",
    "*.tests.*",
    "testing",
    "testing.*",
    "*.testing",
    "*.testing.*",
    "sculptor.testing",
    "sculptor.testing.*",
]

# The next two stanzas ensure that these non-python files are included inside the sculptor wheel.
[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
sculptor = [
    "services/environment_service/providers/docker/imbue_addons/*",
    "services/environment_service/providers/docker/default_devcontainer/*",
    "cli/ssh_config/*",
]
# Sadly, the build system will not complain if the above are misspelled.

[dependency-groups]
dev = [
    "boto3-stubs>=1.40.49",
    "filelock>=3.18.0",
    "moto[s3]",
    "playwright==1.53.0",
    "pre-commit",
    "pyre-check",
    "pytest-cov>=6.2.1",
    "pytest-metadata==3.1.1",
    "pytest-performancetotal==0.2.11",
    "pytest-playwright==0.7.0",
    "pytest-timeout",
    "pytest-xdist>=3.8.0",
    "pytest>=8.4.1",
    "selenium>=4.0.0",
    "syrupy",
    "tabulate>=0.9.0",
    "ty>=0.0.4",
    "websocket-client>=1.5.1",
]

[project.optional-dependencies]
packaging = [
    # When we run pyinstaller, we want to exclude dev dependencies for a smaller executable
    # but pyinstaller itself needs to be available so it can't be a dev dependency.
    "pyinstaller>=6.15.0",
]

[project.scripts]
sculptor_npm_run = "sculptor.scripts.npm_run:app"
sculptor = "sculptor.cli.main:entrypoint"
sculptor_eslint_preprocess = "sculptor.scripts.eslint_preprocess:main"
dev = "sculptor.scripts.dev:app"  # Utilities for development of Sculptor
builder = "builder.cli:app"  # Utilities for building and releasing Sculptor
manual_precommit = "sculptor.scripts.dev:manual_precommit"
cli = "sculptor.cli.dev:typer_cli"  # Command-line utilities of Sculptor per-se Sculptor


[tool.ty.rules]
possibly-missing-attribute = "ignore"

[tool.ty.src]
exclude = [
        "tests/"  # NOTE: We've disabled typechecking for tests for now they descend to base mockery, which is harder to
                  # typecheck effectively.
]
